requirements.txt
ipykernel
flask
requests
python-dotenv
datetime